Start work with smart component
