+++ src/view/popup-close-button-view.js
@@ -7,6 +7,7 @@ const createPopupCloseButtonTemplate = () => (
 );
 
 export default class PopupCloseButtonView extends AbstractView {
+
   get template() {
     return createPopupCloseButtonTemplate();
   }
