--- src/view/popup-comment-view.js
@@ -1,31 +1,0 @@
-import AbstractView from './abstract-view.js';
-import dayjs from 'dayjs';
-
-const createPopupCommentTemplate = (comment) => (
-  `<li class="film-details__comment">
-    <span class="film-details__comment-emoji">
-      <img src="./images/emoji/${comment.emotion}.png" width="55" height="55" alt="emoji-smile">
-    </span>
-    <div>
-      <p class="film-details__comment-text">${comment.comment}</p>
-      <p class="film-details__comment-info">
-        <span class="film-details__comment-author">${comment.author}</span>
-        <span class="film-details__comment-day">${dayjs(comment.date).format('DD/MM/YYYY hh:mm')}</span>
-        <button class="film-details__comment-delete">Delete</button>
-      </p>
-    </div>
-  </li>`
-);
-
-export default class PopupCommentView extends AbstractView {
-  #comment = null;
-
-  constructor(comment) {
-    super();
-    this.#comment = comment;
-  }
-
-  get template() {
-    return createPopupCommentTemplate(this.#comment);
-  }
-}
