+++ src/view/smart-view.js
@@ -1,0 +1,32 @@
+//это файл сделан, при переделывании он понадобится, но сейчас не используется
+import AbstractView from './abstract-view.js';
+
+export default class SmartView extends AbstractView {
+  _data = {};
+
+  updateData = (update) => {
+    if (!update) {
+      return;
+    }
+
+    this._data = {...this._data, ...update};
+
+    this.updateElement();
+  }
+
+  updateElement = () => {
+    const prevElement = this.element;
+    const parent = prevElement.parentElement;
+    this.removeElement();
+
+    const newElement = this.element;
+
+    parent.replaceChild(newElement, prevElement);
+
+    this.restoreHandlers();
+  }
+
+  restoreHandlers = () => {
+    throw new Error('Abstract method not implemented: restoreHandlers');
+  }
+}
